feat(web): split workflow data helpers [OPE-339]
